Speculative fix for a device wrongly enters tablet mode when reopen lid
I cannot repro on my eve or caroline device, but according to comment#4
in bug 1050195, and comment#9 in bug 1050330, it seems theoretically it
can be caused by aync reading of the lid angle driver
"cros-ec-lid-angle". So when the device first starts or comes out from
sleep, the driver "cros-ec-lid-angle" is not present yet, so Chrome is
doing its own lid angle calculation, and later when the reading result
comes back and"cros-ec-lid-angle" flag is properly set, Chrome does not
do the lid angle calculation anymore, but the old lid angle is not reset
to 0.f, so it might interfere with the tablet state calculation.

Bug: 1050195, 1050330
Change-Id: I9c650ee2becbabfb20101a192ebfa5eeb7b677dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2153899
Commit-Queue: Xiaoqian Dai <xdai@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760553}