Caption bubble positioned according to spec.
The bubble is by default centered in its anchor view, and can
be moved anywhere within the anchor view as long as it is not
too close to the sides or bottom. If the bubble is moved within
the anchor view, it tries to stay at the same place with
respect to the anchor as the anchor changes size.

Bug: 1055150
Change-Id: Ic161a99e75bfcc7387b44ee14bed0bc996276e70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2150691
Commit-Queue: Katie Dektar <katie@chromium.org>
Reviewed-by: Abigail Klein <abigailbklein@google.com>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760558}