Fix flakiness in accessibility hit testing browser tests
The cause of the flakiness is that we weren't waiting long enough for
the pinch zoom change to be reflected in the accessibility tree.
Waiting for kLocationChanged appears to address the issue - with this
change I was able to run the flaky test case 1000x without a single
failure.

Bug: 1069850, 1070647
Change-Id: If0cba1804776125885db3d0a2aedfeeca9340048
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2155077
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#760577}