Fetch: Unify response initialization code.
Currently FetchResponseData is initialized from a ResourceResponse in
two places; fetch_manager.cc and fetch_event.cc.  The code in these
two different locations is often out of sync.  This results in the
service worker navigation preload Response not working like other
Response objects.

This CL refactors the initialization into a common routine that is
called from both sites.  This is in preparation to add more attributes
that need to be set during initialization in a later CL.

Bug: 1069813
Change-Id: I67b244bce8066a7e7d291c78fbf48ca863c19598
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2147193
Commit-Queue: Ben Kelly <wanderview@chromium.org>
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760580}