Change Reader Mode to only run on SECURE pages for experiment on Desktop.
HTTP pages can change from WARNING to DANGEROUS if you interact with a
form. There is some chance that a Reader Mode page could load a form, and
Reader Mode doesn't have a good way to convey this security state change,
so limiting Reader Mode to SECURE pages avoids this case.

It is hard to communicate to users the difference between SECURE,
WARNING and Reader Mode in a single security chip, so it is easier to
ensure that pages will be SECURE. (A future CL will avoid security
state changes for SECURE pages by disallowing loading of insecure
resources.)

Also allows Android reader mode to continue running on all pages
as it has done already (feature launched on Android). Android was
restricted to not run on DANGEROUS pages in
https://chromium-review.googlesource.com/c/chromium/src/+/2111075,
but that should only have applied to desktop.

Bug: 1069907
Change-Id: I7895e06132508e5cd37509d890140de167f25f7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2146077
Commit-Queue: Katie Dektar <katie@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Aran Gilman <gilmanmh@google.com>
Reviewed-by: Emily Stark <estark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760602}