[cc/metrics] Use std::bitset instead of base::flat_set.
The CompositorFrameReportingController keeps track of the existing
interactions in a base::flat_set. However, there are only a small number
of such interactions possible at the same time, and a std::bitset is
sufficient to track all of them. So use the more efficient data type.

This makes it possible to simply copy the bitset for each
CompositorFrameReporter instance, rather than keeping a raw pointer back
to the set owned by the controller. This also avoids a data-race when
the reporting of the metrics is moved to a separate thread (in
crrev.com/2135418).

BUG=1072740

Change-Id: If325dbcb9846c2de0c7bbc884585596926c64970
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2156106
Commit-Queue: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Behdad Bakhshinategh <behdadb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760608}