Reland "Add live details, images, duration fields for episodes."
This is a reland of https://crrev.com/c/2148247.

The time in Windows is slightly off. For now, change the test just to
check that the information is printed, not check the exact time value.
This is an experimental page so we can relax the test to unblock further
changes while we debug this.

Bug: 1065986
Change-Id: I99df2fb2e11f183687bb73477f7253f8ce09f5b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2148247
Commit-Queue: Sam Bowen <sgbowen@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760284}

TBR=beccahughes@chromium.org, dcheng@chromium.org

Change-Id: I99df2fb2e11f183687bb73477f7253f8ce09f5b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2156610
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Commit-Queue: Sam Bowen <sgbowen@google.com>
Cr-Commit-Position: refs/heads/master@{#760619}