SubtreeVisibility: Don't update layout on focus checks.
This patch moves the style and layout tree update earlier in the
focused style check since is needed to check if the element has a
layout object. It also removes the layout update, which isn't needed
for this function. If a caller depends on the layout being up to date,
then it should be the caller that updates it. However, I don't think
there are any callers that currently depend on this.

R=chrishtr@chromium.org, rakina@chromium.org, aleventhal@chromium.org

Change-Id: Ic6c6bef6602cbe15111bd282e9ae69fd700dd0ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2155005
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: vmpstr <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760632}