Completes browser side support for subframe translation behind a flag
This is a subset (3rd of 3) of a large change proven out in CL 2117459.
Its focus is to maintain the existing ContentTranslateDriver code as much
as possible for patch-ability while adding a PerFrameContentTranslateDriver
that is meant to replace it in the future.

[Update: there will also be a follow-up CL to move the CLD3 language
detection work into a sandboxed utility process.]

The overall change description:

This conditionally will perform language detection in the browser
process using Accessibility apis to get page contents in order to
determine language from subframes as well as main frame and then
supports translating subframes in addition to the main frame. This
new behavior is available when Feature "TranslateSubFrames" is
enabled; otherwise, the existing language detection (using
WebFrameContentDumper) and main frame translation support is preserved
(in the spirit of the current guidelines against refactors).

Issue 1064974 tracks follow-up clean-up tasks for the duplicated code
in this CL once the feature is launched (once refactors encouraged
again).

Bug: 797328, 1063520
Change-Id: I47747e6a49d5c5568de86ed1a1189c27b49ad8db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2148128
Reviewed-by: anthonyvd <anthonyvd@chromium.org>
Reviewed-by: Megan Jablonski <megjablon@chromium.org>
Commit-Queue: Doug Arnett <dougarnett@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760711}