Replace NIK in ResourceRequest::TrustedParams with IsolationInfo.
IsolationInfo includes the SiteForCookies, which must be consistent
with the NIK, and a rule for updating both values in the case of
redirects. This will ensure what were previously 4 values
(SiteForCookies, NIK, and how to update each) are all consistent, and
will allow us to migrate off of having untrusted renderer processes set
the SiteForCookies.

Currently the IsolationInfo's SiteForCookies is ignored, apart from
DCHECKing that it's consistent with the ResourceRequest's
SiteForCookies. We'll switch over to using it, and remove the legacy
value, once they're always consistent.

BUG=1060631

Change-Id: I14205172c041c3dc7bdf911044ccff9dbb85bbcc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2142742
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Joshua Bell <jsbell@chromium.org>
Reviewed-by: Roger McFarlane <rogerm@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Peter Kotwicz <pkotwicz@chromium.org>
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Min Qin <qinmin@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Benoit L <lizeb@chromium.org>
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Commit-Queue: Becca Hughes <beccahughes@chromium.org>
Auto-Submit: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760717}