Add the "is declarative shadow root" flag, and set it appropriately
This CL implements the "is declarative shadow root" flag, and makes
sure it gets set or cleared appropriately.

This flag will be used in a followup CL, which modifies the behavior
of attachShadow() to allow "overriding" a declarative shadow root.
But in this CL, expect no behavior changes.

Bug: 1042130
Change-Id: I335df00967c2fa9dcce4ded8f6a7cb208a901e68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2155995
Commit-Queue: Mason Freed <masonfreed@chromium.org>
Auto-Submit: Mason Freed <masonfreed@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760720}