appcache: only store token_expires on cache
This reverts some previous work here to clean up code that will become
unused.  Sorry to go back and forth on this, but it wasn't clear what
was needed until I understood this more and got to the end.

The current state with this patch:
* only store token_expires on the cache (and cache record) itself
* remove code storing it on entries, groups, namespaces
* leave the database code as-is with the extra (unused) fields

Future patches:
* read token_expires from manifest, not headers (headers are cumbersome)
* AppCacheDatabase::FindCache can be used to examine token_expires and
  cut off any unwanted AppCache usage

Bug: 582750
Change-Id: I155ecf122886566f30a5be38b96994558a59d3cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2157602
Commit-Queue: Victor Costan <pwnall@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760729}