Change offscreen back texture initial size to 64x64
There was a bug on kukui where this code path would get hit and fail.
The minigbm allocation would succeed, but then attaching it to the FB
would fail because the mediatek DRM driver has a minimum width/height
requirement of 64 for the framebuffer. This buffer is just throwaway
and gets reallocated after the offscreen buffer gets swapped, so
allocating a slightly larger one should cause no problems and the
small increase wastes a negligible amount of memory.

This bug specifically manifested itself in Zoom in Chrome on kukui.

BUG=b:151774454
TEST=Zoom video renders on krane

Change-Id: I068273d971eb08a5ff7d5856ddfa4d9b8854b0b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2155381
Commit-Queue: J Kardatzke <jkardatzke@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760744}