[CrOS Settings] Fix crash in AccessibilityMainHandler
This crash was triggered when the handler tried to notify JS about a
change in screen reader state while JS was disallowed.

The fix is to ensure that the handler for screen reader state changes is
only added when JS is allowed in the first place. Note that the
"getScreenReaderState" message is sent to the handler on page load, so
we are still guaranteed that the change listener will be added when the
page loads.

Fixed: 1014164
Change-Id: Ibed22468e731bd8b0a83d14335a0947fda1cb26f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2157826
Reviewed-by: dpapad <dpapad@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760745}