Allow embedders to specify additional providers
Some providers represent core functionality or are relatively small.
These are the orientation providers and the ability to plug in to the
isolated_xr_service (which is desktop only). However, other providers,
especially the android specific ones, can be quite large and it doesn't
really make sense to support them on all platforms (e.g. gvr). In order
to avoid binary size growth on the 64 bit images, this change allows
embedders to specify additional providers (beyond the default ones),
to be linked in.

Bug: 1070807
Change-Id: I238eb495581dfde1af8ab2633d5b70357165ce23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2154736
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Klaus Weidner <klausw@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760772}