[File System] Fix bug in logic for writing to sandboxed files.
SandboxFileStreamWriter assumed that if a write operation finishes
synchronously, it must have failed. However with the in-memory file
system in incognito mode it is also possible for writes to succeed
synchronously. This fixes the logic to update all the correct state
if that happens.

Bug: 1068360
Change-Id: Ia31ae56be407e1561dc65f562caf0010b1b17a55
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2155272
Commit-Queue: Victor Costan <pwnall@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760783}