[MIX-DL] Add Android UI via Android infobar.
This CL extends support of Mixed Content Download Blocking to Android.
It adds a new infobar that almost exactly mirrors the Dangerous Download
infobar.

This infobar supports both the WARN and BLOCK states that Desktop MIX-DL
blocking supports, but the only difference between WARN and BLOCK are
the button ordering/defaults. In WARN, the default action is Keep, while
in BLOCK, the default action is Discard.

Both variants use the standard blue warning triangle logo. This is
consistent with the Dangerous Download infobar.

This UI is hidden behind the same flag as the rest of MIX-DL.

BUG=960819

Change-Id: I7f3c8f2c7b2bd8871c01d282b1fe4863ce53ea0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2155400
Reviewed-by: Xing Liu <xingliu@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Auto-Submit: Joe DeBlasio <jdeblasio@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760815}