[ios] Handle EditBookmarks pref in ui/activity_services/
EditBookmarksEnabled is an enterprise policy that determines whether
the user can edit any bookmarks. The policy is mapped to a user pref.

This CL is part of a series of CLs that add support for this
enterprise policy on the iOS platform.

Specifically in this CL, the add/edit bookmark share activity is
removed when EditBookmarksEnabled is false.

There is no good way to unit test items in the share menu. And eg tests
can't interact with the share menu yet.

Bug: 1072326
Change-Id: If0455b9878042f33ba1d9f72be1302394b8a92f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2156141
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760828}