🤝 Create Running in Chrome Snackbar.
This CL creates the alternative Snackbar that will replace the existing
Infobar in the new design. (A Snackbar dismisses itself after some time
while an Infobar does not.)

It also sets up a bunch of infrastructure for the remainder of the
changes - adding tests for the new and legacy behaviors, creating a
feature flag and a documentation markdown file to explain what's going
on.

Bug: 1068106
Change-Id: I71b47ac447ff4b5c38873071402d56780b3c81c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2134304
Reviewed-by: Yusuf Ozuysal <yusufo@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Commit-Queue: Peter Conn <peconn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760830}