[@property] Align CSSOM behavior with spec
This CL aligns the CSSOM API for @property rules with the spec. (The
current CSSOM behavior for @property in Blink predates the
CSSOM/@property chapters in the spec).

There are two main changes:

 - Expose name/syntax/inherits/initialValue attributes instead of a
   generic CSSStyleDeclaration.
 - Implement cssText approximately* according to spec.

The tests in fast/css will be moved to WPT separately, when I'm sure
they're fully updated to match the spec.

* I made several mistakes in the spec: 1) initialValue instead of
  initial-value, 2) adding whitespace after the "initial-value:"
  serialization (doesn't round-trip, since whitespace is significant
  here), 3) doesn't properly handle missing descriptors.

Bug: 978783, 973830
Change-Id: I1dff26638f24ac9c7973652240bc7a4bcb3ff6ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2152801
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760845}