[Android][TouchToFill] Check and recreate Android view before usage.
Relying on GetWindowAndroid() returning a non-null value causes crashes
in various instances. With this CL, the window is null-checked before
use an creation is deferred to when it's actually used.
If creation doesn't succeed on first try (e.g. because the window isn't
yet attached), a consecutive call might succeed.

Bug: 1051033
Change-Id: I8922c86012bf5f2f1f14ad39c804702e70093ba6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2156885
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760871}