Optimize some data deletion code
This CL does a few things:
1. Currently, both the StoragePartition ClearData and the
BrowsingDataRemover use Origins internally. However, the
BrowsingDataRemover exposes a generic matching interface that only
accepts GURLs.

This leads to lots of wasteful conversions. For every internal DB check
to clear data for an origin in StoragePartition, we need to convert the
origin into a GURL, and then back into an Origin again. Both of those
operations incur expensive allocations because Origin and GURL use
a different internal representation.

This CL adds a browsing data filter that accepts Origins, in addition
to GURLs, and uses it in the StoragePartition.

2. GetDomainAndRegistry has no variants that accept an Origin, and the
one that accepts a host does an expensive canonicalization step
(more allocations). This CL adds a new method to get the domain and
registry from an Origin which will have already canonicalized its host
internally (similar to the GURL variant).

3. Minor cleanups in DoesOriginMatchMaskAndPredicate to re-order conditionals
to avoid expensive Origin::GetURL computations.

This CL has no intended behavior changes.

Bug: None
Change-Id: I7964b896479aab2bf01e8247e562bd55dff18927
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2151840
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760943}