browser: Add Get{Vm,ArcSystem}ProcessList traces
While profiling Chrome on CrOS, we noticed a significant amount of time is
spent in Get{Vm,ArcSystem}ProcessList even if the task manager is not
running.

On a nocturne m3 we observed 2% total of the CPU used is spent in these
two functions.
This CL adds traces for those two functions.
Capturing a trace confirms that we call those two functions every few seconds
and they take about 80ms of CPU time each.

Bug: b/154362057
Change-Id: Ie851643dc1093e400282aeb976c880b9901f3138
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2153827
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Daniele Castagna <dcastagna@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760949}