Specify logic in tests to wait for TabGridDialog show/hide animation
Currently, we use TabUiTestHelper#isShowingPopupTabList to verify if
TabGridDialog is showing or hidden. However, even if this call is
good enough to verify that the dialog is completely hidden, passing
this check doesn't ensure that the dialog showing animation has ended.
This is a problem exposed by extra checks introduced in
http://crrev.com/c/2143671. This CL fixes this issue by separating the
logics of checking dialog showing and hiding into two methods with
different criteria.

Bug: 1072376, 1072374
Change-Id: Iabb1e9488a8f94ced36988b90b08ec90df2ba911
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2158054
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Commit-Queue: Yue Zhang <yuezhanggg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760967}