Split out method for processing keyframes rule when creating a CSS animation
This patch moves the processing of the keyframes rule into a separate
method as a preliminary step in aligning the keyframe generate process
for CSS animations with the spec
(https://drafts.csswg.org/css-animations-2/#keyframes).  Once this
change is in, further refinement to align the algorithm with the spec
should be easier to follow.


Bug: 1070627
Change-Id: I04c558afa8f3aa9db85c8c473e4fbc5577633da8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2154824
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/master@{#760981}