[Paint Preview] Restrict capture size over mojo
This CL adds support for plumbing a maximum capture size restriction
to the PaintPreviewRecorde returning the size of the output.

For now, TabbedPaintPreviewPlayer will use this to cap per SkPicture
capture size to 5 MB.

In a follow up the returned size information will be used to restrict
the total capture size rather than the size per renderer.

Bug: 1071446
Change-Id: Ib809d3156c11e6a8ee6144663794fe261f89f635
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2152957
Reviewed-by: Mehran Mahmoudi <mahmoudi@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761053}