bluetooth: clusterfuzz: Fix PRESUBMIT script
This change fixes an error when running the PRESUBMIT script in
//third_party/blink/renderer/modules/bluetooth/testing/clusterfuzz. This
error was caused by attempting to replace a substring in a UTF-8 encoded
string with an ASCII encoded string. The fix converts the ASCII string
to UTF-8 prior to this replacement operation.

Bug: 1064538
Change-Id: I2c69ed4dc7bb9e47a2b0ed3150c90c6357df626d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2149891
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Ovidio de Jesús Ruiz-Henríquez <odejesush@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761057}