[@property] Determine the registration
The current behavior for dealing with multiple registrations on the
same custom property was highly temporary, and not based on any spec.
The spec has since been updated to describe how it should work:
in short, registrations originating from CSS.registerProperty are
stored independently from those originating from @property, and
CSS.registerProperty-registrations (if present) win over @property-
registrations.

Changes in this CL:

 - Separates storage for CSS.registerProperty-registrations
   (registered_properties_) and @property-registrations
   (declared_properties_). The old code put @property registrations
   in the same map as CSS.registerProperty-registrations, and only
   if no registration exited already. (Meaning for example that the
   first @property seen would win---opposite of how it's spec'd to
   work now).
 - Determines the registration [1] correctly. I.e. first checks the
   the CSS.registerProperty-registrations, and the @property ones.
 - Adds an iterator which yields the currently active registrations
   across both registration types. (I wanted to do this separately,
   but initial values won't work properly without it, and there's a
   test that relies heavily on that being functional).

Note that it's not explicitly stated in the spec that the storage for
the two registration types shall be separate, but that _is_ the
background for the current spec text on the subject. The rationale is
that it makes it easier to reason about which registration is the
current one when @property rules can appear and disappear via JS.
It's also expected that CSS.unregisterProperty will be added in the
future, making the situation even more unstable.

[1] https://drafts.css-houdini.org/css-properties-values-api-1/#determining-registration

Bug: 973830
Change-Id: I69b48629034b333ebed5d374f943d08b39579334
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2157071
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#761075}